Provide direct access to intermediates and hashes
OVERVIEW

The `v2` series of releases is a big change in focus for the project.

Instead of providing an opaque `*x509.CertPool` and a
`VerifyConnection` callback for use with a `tls.Config` for a client
connection, this series of releases drops that support and provides
direct access to collections of intermediate certificates and hashes
for direct use by clients.

The `v2` series provides:

- valid intermediate certificates and expired intermediate
  certificates but not revoked intermediate certificates which chain
  to root certificates in Mozilla's program
- known unexpired, unrevoked intermediate certificates chaining to
  roots with Websites trust in the Mozilla Root Program
- revoked intermediate CA certificates which chained to roots in the
  Mozilla Root Program
- certificate hashes for all sets

The audience for this functionality is primarily diagnostic tools
which need direct access to `x509.Certificate` values for certificate
chain analysis purposes.

CHANGES

- drop opaque `*x509.CertPool`
- drop `VerifyConnection` callback
  - used with a `tls.Config` for a client connection
- provide direct access to collections of intermediate certificates
  and hashes via `MustGet*` (panics if unsuccessful) and `Get*`
  (returns error if unsuccessful) functions
  - existing known unexpired, unrevoked intermediate certificates
    chaining to roots with Websites trust in the Mozilla Root Program
  - new valid intermediate certificates and expired intermediate
    certificates but not revoked intermediate certificates which chain
    to root certificates in Mozilla's program
  - new revoked intermediate CA certificates which chained to roots in
    the Mozilla Root Program
  - new certificate hashes for all provided intermediate cert
    collections
- bundle copies of upstream intermediate reports (CSV format)
- rename existing `intermediates.pem` file to match upstream report
  name
- bundle generated PEM format files
- bundle generated text files of intermediate hashes
- update releases to provide downloaded and generated files
- add upstream license information for CCADB data
- refresh documentation to reflect changes

Other smaller changes have also been applied to support these primary
changes.

CREDIT

The following LLM sources were used for inspiration or starting code
samples for some of the included changes:

- ChatGPT, OpenAI
- Google Gemini
- Claude (Anthropic AI assistant)

All were helpful in pointing me in the right direction when I needed
it, particularly function naming patterns.

REFERENCES

- refs GH-2